Fix homepage links not opening in new tab
Removing the 'href' attribute of the anchor tags (and opening the links
using js after a 'click' event) made it unable to open the links in a
new tab. The was done to hide the 'overLink' status panel which Firefox
(and most other browsers) show by default when hovering a link. This
information is non-informative (cryptic url) and useless to the user.

Remove the `hideInternalLinksInfoWhenReady()` and sub-functions. This
reverts the 'hiding' feature, but restores the ability to open links
in a new tab (which is a more essential feature).